// src/client/models/Page_User_JQQ_.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { User_JQQ } from './User_JQQ';

export type Page_User_JQQ_ = {
    items: Array<User_JQQ>;
    total: number;
    page?: number;
    size?: number;
    pages?: number;
};