// src/client/models/Page_UserOut2Schema_.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { UserOut2Schema } from './UserOut2Schema';

export type Page_UserOut2Schema_ = {
    items: Array<UserOut2Schema>;
    total: number;
    page?: number;
    size?: number;
    pages?: number;
};